requirements.txt
Django==1.9
MySQL-python==1.2.5
argparse==1.2.1
django-cors-headers==1.1.0
django-reroute==1.1.1
wsgiref==0.1.2
